refactor: the MLX model download moves into one type
Two models are fetched the same way — the lock, the staging directory, the
file list, the de-duplicated progress. WordVectors held all of it. The
sentence readings need it next, so it becomes MLXModelCache and the
FolderTokenizer adapter beside it.

No behaviour change: same repository, same files, same cache directory,
same lock path.
